requirements.txt
langchain
langchain-community
langchain-openai
langchain-deepseek
langchain-ollama
os 
pandas 
xml.etree.ElementTree
langchain-core
python-dotenv